backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pdfplumber>=0.11.0
sentence-transformers>=2.7.0
numpy>=1.26.0
google-genai>=1.0.0
python-dotenv>=1.0.0
